+++ releases/ChangeLog.md
@@ -13,6 +13,7 @@
   private key corresponding to the address specified in HSM policy. Attestation signature
   MUST be provided in PSBT in a new proprietary field. Thanks to
   [@straylight-orbit](https://github.com/straylight-orbit) for this powerful new feature.
+- Enhancement: Ephemeral Seeds: Advanced/Tools -> Ephemeral Seed (more info in `docs/ephemeral.md`)
 - Bugfix: Correct parsing of unknown fields in PSBT: they are now passed through.
 - Bugfix: Share single address over NFC from address explorer menu.
 - Bugfix: Using lots of trick pins (7+), could lead to a case where the Coldcard would